Adding and Editing the feed is almost fully functional
Holding a turn will be delayed for a while. It's only really used with preparing an action and such. I would like to clean up my JScript at some point in the future, but that is in the backlog. Next thing to do is to allow characters to be removed or added in the encounter.